Extract Signature to separate file
